// src/types/generated/types.gen.ts
export type Idea = {
  id: string;
  title: string;
  shortDescription: string;
  description: string;
  tags: Array<string>;
  authorId: string;
  author: UserSummary;
  reviewsCount: number;
  avgRating: number | null;
  createdAt: string;
  updatedAt: string;
};

export type UserSummary = {
  id: string;
  email: string;
  username: string;
};
// ...
export type CreateIdeaData = {
  /**
   * Idea data
   */
  body: {
    title: string;
    shortDescription: string;
    description: string;
    tags?: Array<string>;
  };
  path?: never;
  query?: never;
  url: '/ideas';
};