// src/types/generated/zod.gen.ts
export const zCreateIdeaBody = z.object({
  title: z.string().min(1).max(255),
  shortDescription: z.string().min(1).max(500),
  description: z.string().min(1),
  tags: z.array(z.string()).optional().default([]),
});